!!!!starting to notice something weird!!!
one of the papers states that the expected critical state of a bounded
1D sandpile looks kind of like a triangle. BUT my simulations did not
observe this. instead the pile fills up, regardless of how I attempt to
distribute the addition of new sand.
I’ve made a simulation that uses some slope related metric to determine
whether or not the pile topples, which does start to exhibit the
expected critical state. except that I think this is wrong. attempting
to port this to the 2D simulation and see whether or not we lose
features like fractals, which would suggest that using slope does not
work.
foreseen issues: there’s two ways to approach the slope. One is to use
overall difference, which seems less realistic/intuitive. the other is
just to only dump only in the direction where slope exceeds critical
value K. not sure what happens in the first case, but i’m suspect that
the second one will no longer exhibit self-organized criticality.